feat: added and implemented spinner loading
