+++ src/components/Spinner/Spinner.js
@@ -1,0 +1,9 @@
+import { Overlay, Loader } from './Spinner.styled.jsx'
+
+const Spinner = () => (
+	<Overlay>
+		<Loader />
+	</Overlay>
+)
+
+export default Spinner
